refactor 'api' sub files
