+++ magpie/api/exception.py
