+++ magpie/api/login/__init__.py
@@ -3,7 +3,7 @@ LOGGER = get_logger(__name__)
 
 
 def includeme(config):
-    from magpie.api import api_rest_schemas as s
+    from magpie.api import schemas as s
     LOGGER.info("Adding api login...")
     # Add all the rest api routes
     config.add_route(**s.service_api_route_info(s.SessionAPI))
