+++ magpie/api/management/resource/resource_views.py
@@ -1,4 +1,4 @@
-from magpie.api import api_requests as ar, api_except as ax, api_rest_schemas as s
+from magpie.api import requests as ar, exception as ax, schemas as s
 from magpie.api.management.service.service_utils import get_services_by_type
 from magpie.api.management.service.service_formats import format_service_resources
 from magpie.api.management.resource import resource_utils as ru, resource_formats as rf
