+++ magpie/api/management/service/__init__.py
@@ -1,4 +1,4 @@
-from magpie.api import api_rest_schemas as s
+from magpie.api import schemas as s
 from magpie.utils import get_logger
 LOGGER = get_logger(__name__)
