+++ magpie/api/schemas.py
