+++ magpie/definitions/__init__.py
@@ -2,4 +2,4 @@
 def includeme(config):
     from magpie.utils import get_logger
     logger = get_logger(__name__)
-    logger.info('Adding definitions...')
+    logger.info("Adding definitions...")
